requirements.txt
absl-py==0.12.0
aiobotocore==2.3.2
aiohttp==3.8.1
aioitertools==0.8.0
aiosignal==1.2.0
alembic
altair==4.2.0
anyio==3.4.0
apache-beam==2.39.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astor==0.8.1
astroid==2.7.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autopage==0.5.1
autopep8==1.6.0
avro-python3==1.10.2
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
black==21.12b0
bleach==4.1.0
blinker==1.4
botocore==1.24.21
boto3
branca==0.4.2
cachetools==4.2.4
category-encoders==2.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==2.0.0
cmaes==0.8.2
cmd2
colorama==0.4.4
colorcet==3.0.0
colorlog==6.7.0
commonmark==0.9.1
contextlib2==21.6.0
coverage==6.2
crcmod==1.7
cryptography==37.0.2
cycler==0.11.0
Cython==0.29.26
dask==2021.12.0
databricks-cli==0.16.2
deap==1.3.1
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.1.1
docker==5.0.3
docopt==0.6.2
docutils==0.18.1
entrypoints==0.3
et-xmlfile==1.1.0
fastapi==0.70.1
fastavro==1.4.12
flake8
Flask==2.0.2
Flask-Cors==3.0.10
flatbuffers==2.0
folium==0.12.1.post1
fonttools==4.28.5
formulaic==0.3.4
frozenlist==1.2.0
fsspec==2022.5.0
future==0.18.2
gast==0.4.0
gcsfs==2022.5.0
gensim==4.1.2
geographiclib==1.52
geopy==2.2.0
gitdb==4.0.9
GitPython==3.1.24
google-api-core==2.3.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.31.0
google-cloud-bigquery-storage==2.10.1
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
google-trans-new==1.1.9
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.6.0
hdfs==2.7.0
htmlmin==0.1.12
httplib2==0.19.1
idna==3.3
ImageHash==4.2.1
imageio==2.13.5
imbalanced-learn==0.8.1
imblearn==0.0
importlib-metadata==4.10.0
importlib-resources==5.4.0
iniconfig==1.1.1
install==1.3.5
interface-meta==1.3.0
ipdb==0.13.9
ipykernel==6.6.1
ipympl==0.8.5
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
isoweek==1.3.3
itsdangerous==2.0.1
jedi==0.18.1
jeepney==0.8.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.0.1
json5==0.9.6
jsonschema==4.3.3
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-resource-usage==0.6.1
jupyter-server==1.13.1
jupyter-server-mathjax==0.2.3
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.2
jupyterlab-widgets==1.0.2
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
keyring==23.5.1
kiwisolver==1.3.2
kneed==0.7.0
lazy-object-proxy==1.7.1
libclang==12.0.0
libcst==0.3.23
locket==0.2.1
lvis==0.5.3
lxml==4.7.1
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
memoized-property==1.0.3
missingno==0.5.0
mistune==0.8.4
mlflow
multidict==5.2.0
multimethod==1.6
mypy-extensions==0.4.3
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.4.0
nbdime==3.1.1
nbformat==5.1.3
nbimporter==0.3.4
nbresult==0.0.4
nest-asyncio==1.5.4
networkx==2.6.3
nltk==3.6.7
notebook==6.4.6
numpy
oauthlib==3.1.1
onedrivesdk
opencv-python==4.5.5.64
openpyxl==3.0.10
opt-einsum==3.3.0
optuna
orjson==3.6.8
packaging==21.3
pandas
pandas-gbq==0.16.0
pandas-profiling==3.1.0
pandocfilters==1.5.0
param==1.12.2
parso==0.8.3
partd==1.2.0
pathspec==0.9.0
patsy==0.5.2
pbr==5.10.0
pexpect==4.8.0
phik==0.12.0
pickleshare==0.7.5
Pillow==9.0.0
pkginfo==1.8.2
platformdirs==2.4.1
plotly==5.5.0
pluggy==1.0.0
pmdarima==1.8.4
prettytable==3.4.1
prometheus-client==0.12.0
prometheus-flask-exporter==0.18.7
promise==2.3
prompt-toolkit==3.0.24
property-cached==1.6.4
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.11.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycountry==22.3.5
pycparser==2.21
pyct==0.4.8
pydantic==1.9.0
pydata-google-auth==1.3.0
pydeck==0.7.1
pyDeprecate==0.3.2
pydot==1.4.2
pyflakes==2.4.0
pygeohash==1.2.0
Pygments==2.11.2
pyhdfe==0.1.0
pylint==2.10.2
pymongo==3.12.3
Pympler==1.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-editor==1.0.4
python-gitlab==3.0.0
python-multipart==0.0.5
python-slugify==5.0.2
pytorch-forecasting==0.10.3
pytorch-lightning==1.7.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
querystring-parser==1.2.4
readme-renderer==35.0
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.1
rsa==4.8
s3fs==2022.5.0
scikit-image==0.19.1
scikit-learn
scipy
seaborn==0.11.2
SecretStorage==3.3.2
semver==2.13.0
Send2Trash==1.8.0
setuptools-scm==6.4.2
six==1.14.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
sqlparse==0.4.2
starlette==0.16.0
statsmodels==0.13.1
stellargraph==1.2.1
stevedore
stopit==1.1.2
streamlit==1.9.1
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
tensorflow-metadata==1.5.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
tf-slim==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
torch==1.12.1
torchmetrics==0.9.3
tornado==6.1
TPOT==0.11.7
tqdm==4.62.3
traitlets==5.1.1
twine==4.0.0
typing-inspect==0.7.1
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
update-checker==0.18.0
urllib3==1.26.7
uvicorn==0.16.0
validators==0.19.0
visions==0.7.4
watchdog==2.1.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.12.1
xgboost==1.5.1
XlsxWriter==3.0.2
yapf==0.32.0
yarl==1.7.2
zipp==3.7.0